+++ packages/contentful--app-scripts/src/clean-up/index.ts
@@ -1,24 +1,7 @@
 import { getCleanUpSettingsArgs } from './get-clean-up-settings';
 import { buildCleanUpSettings } from './build-clean-up-settings';
 import { cleanUpBundles } from './clean-up-bundles';
-import { Organization } from '../organization-api';
-import { Definition } from '../definition-api';
-
-export interface CleanupOptions {
-  organizationId?: string;
-  definitionId?: string;
-  token?: string;
-  host?: string;
-  keep?: string;
-}
-
-export interface CleanupSettings {
-  keep: number;
-  host?: string;
-  accessToken: string;
-  organization: Organization;
-  definition: Definition;
-}
+import { CleanupOptions } from '../types';
 
 const interactive = async (options: CleanupOptions) => {
   const settings = await buildCleanUpSettings(options);
