+++ packages/contentful--app-scripts/src/upload/create-app-upload.test.ts
@@ -2,7 +2,7 @@ import assert from 'assert';
 import { stub, match, SinonStub } from 'sinon';
 import proxyquire from 'proxyquire';
 import { ClientAPI } from 'contentful-management';
-import { UploadSettings } from '.';
+import { UploadSettings } from '../types';
 
 describe('createAppUpload', () => {
   let createAppUpload: typeof import('./create-app-upload').createAppUpload, clientMock: ClientAPI;
