+++ packages/contentful--app-scripts/src/upload/create-app-upload.ts
@@ -3,7 +3,7 @@ import { createZipFileFromDirectory } from './create-zip-from-directory';
 import { validateBundle } from './validate-bundle';
 import { showCreationError } from '../utils';
 import { createClient } from 'contentful-management';
-import { UploadSettings } from '.';
+import { UploadSettings } from '../types';
 
 async function createAppBundleFromFile(orgId: string, token: string, zip: Buffer, host = '') {
   const client = createClient({ accessToken: token, host, hostUpload: host.replace(/^api/i, 'upload') });
@@ -12,10 +12,10 @@ async function createAppBundleFromFile(orgId: string, token: string, zip: Buffer
 }
 
 export async function createAppUpload(settings: UploadSettings) {
-  validateBundle(settings.bundleDirectory || '.');
+  validateBundle(settings.bundleDirectory || '../types');
   let appUpload = null;
   const zipFileSpinner = ora('Preparing your files for upload...').start();
-  const zipFile = await createZipFileFromDirectory(settings.bundleDirectory || '.');
+  const zipFile = await createZipFileFromDirectory(settings.bundleDirectory || '../types');
   zipFileSpinner.stop();
 
   if (!zipFile) return;
