+++ packages/contentful--app-scripts/src/upload/index.ts
@@ -2,32 +2,7 @@ import { activateBundle } from '../activate/activate-bundle';
 import { getUploadSettingsArgs } from './get-upload-settings-args';
 import { createAppBundleFromSettings } from './create-app-bundle';
 import { buildAppUploadSettings } from './build-upload-settings';
-import { Organization } from '../organization-api';
-import { Definition } from '../definition-api';
-import { FunctionAppAction } from '../utils';
-
-export interface UploadOptions {
-  organizationId?: string;
-  definitionId?: string;
-  token?: string;
-  bundleDir?: string;
-  comment?: string;
-  skipActivation?: boolean;
-  host?: string;
-  userAgentApplication?: string;
-}
-
-export interface UploadSettings {
-  organization: Organization;
-  definition: Definition;
-  accessToken: string;
-  comment?: string;
-  bundleDirectory: string;
-  skipActivation?: boolean;
-  userAgentApplication?: string;
-  host?: string;
-  actions?: FunctionAppAction[];
-}
+import { UploadOptions, UploadSettings } from '../types';
 
 async function uploadAndActivate(settings: UploadSettings) {
   const bundle = await createAppBundleFromSettings(settings);
